+++ fabric/src/main/java/io/github/sakurawald/module/mixin/command_event/PlayerListManagerMixin.java
@@ -33,13 +33,13 @@ public class PlayerListManagerMixin {
     {
         var onPlayerJoinedConfig = CommandEventInitializer.config.model().event.on_player_joined;
         if (onPlayerJoinedConfig.enable) {
-            CommandExecutor.execute(ExtendedCommandSource.asConsole(player.getCommandSource()), onPlayerJoinedConfig.command_list);
+            CommandEventInitializer.executeCommandOnEvent(player, onPlayerJoinedConfig.command_list);
         }
 
         var onPlayerFirstJoinedConfig = CommandEventInitializer.config.model().event.on_player_first_joined;
         if (onPlayerFirstJoinedConfig.enable) {
             if (player.getStatHandler().getStat(Stats.CUSTOM.getOrCreateStat(Stats.LEAVE_GAME)) < 1) {
-                CommandExecutor.execute(ExtendedCommandSource.asConsole(player.getCommandSource()), onPlayerFirstJoinedConfig.command_list);
+                CommandEventInitializer.executeCommandOnEvent(player, onPlayerFirstJoinedConfig.command_list);
             }
         }
     }
@@ -54,7 +54,7 @@ public class PlayerListManagerMixin {
         var config = CommandEventInitializer.config.model().event.after_player_respawn;
         if (config.enable) {
             ServerPlayerEntity newPlayer = cir.getReturnValue();
-            CommandExecutor.execute(ExtendedCommandSource.asConsole(newPlayer.getCommandSource()), config.command_list);
+            CommandEventInitializer.executeCommandOnEvent(newPlayer, config.command_list);
         }
     }